project/requirements.txt
Flask==2.3.3
PyMuPDF==1.23.3
Pillow==10.0.0
numpy==1.25.2
pytesseract==0.3.10